refactor: expose structured tools
