+++ tests/test_tools.py
@@ -1,7 +1,7 @@
 import pytest
 from orchestrator import crud
 from orchestrator.models import ProjectCreate, EpicCreate, FeatureCreate, USCreate
-from agents.tools import (
+from agents.handlers import (
     create_item_tool,
     update_item_tool,
     find_item_tool,
